Model and Db connection
I added model and Db connection and also little changed in script file, little correction in html and about html
