+++ about.html
@@ -55,6 +55,7 @@
                         Students are always at our heart we are student centric.
                         We are continuously innovating to enhance the student experience.
                         As we were all international student once we believe in giving back to community.
+                        Hope you have Great Time with Us.
                     </p>
                 </article>
             </div>
