+++ essential.html
@@ -41,23 +41,24 @@
                         warehouse.
                         Dont worry! We have got you all the options for you in the below link.
                         Check out our job listings to find the perfect job for you.
-                        <br>
-                        For Hospitality please refer the link- <a
-                            href="https://au.indeed.com/jobs?q=Hospitality+Jobs+For+Students&l=&from=searchOnHP&vjk=d1cf8caff5eef6f8"
-                            target="_blank">Indeed Jobs</a>
-                        <br>
-                        For Warehouse Jobs please refer the link- <a href="https://sidekicker.com/au/work/"
-                            target="_blank">Warehouse</a>
-                        <br>
-                        For Delivery Jobs please refer the link- <a href="https://dasher.doordash.com/en-au"
-                            target="_blank">Delivery Jobs</a>
-                        <br>
-                        Oz staff Jobs for students please refer the link- <a href="https://www.ozstaff.com/search-job"
-                            target="_blank">Oz Staff</a>
-                        <br>
-                        Retail Jobs for students please refer the link- <a
-                            href="https://au.indeed.com/q-student-retail-l-melbourne-vic-jobs.html?vjk=62821a828f86f56c"
-                            target="_blank">Retail Jobs</a>
+                    </p>
+                    <br>
+                    For Hospitality please refer the link- <a
+                        href="https://au.indeed.com/jobs?q=Hospitality+Jobs+For+Students&l=&from=searchOnHP&vjk=d1cf8caff5eef6f8"
+                        target="_blank">Indeed Jobs</a>
+                    <br>
+                    For Warehouse Jobs please refer the link- <a href="https://sidekicker.com/au/work/"
+                        target="_blank">Warehouse</a>
+                    <br>
+                    For Delivery Jobs please refer the link- <a href="https://dasher.doordash.com/en-au"
+                        target="_blank">Delivery Jobs</a>
+                    <br>
+                    Oz staff Jobs for students please refer the link- <a href="https://www.ozstaff.com/search-job"
+                        target="_blank">Oz Staff</a>
+                    <br>
+                    Retail Jobs for students please refer the link- <a
+                        href="https://au.indeed.com/q-student-retail-l-melbourne-vic-jobs.html?vjk=62821a828f86f56c"
+                        target="_blank">Retail Jobs</a>
                     </p>
                 </article>
                 <br>
@@ -73,12 +74,12 @@
                     </p>
                     <br>
                     For Amber Student please refer the link- <a
-                        href="https://www.amberstudent.com/places/search/melbourne-1811011846366"
-                        target="_blank">Amber Student</a>
+                        href="https://www.amberstudent.com/places/search/melbourne-1811011846366" target="_blank">Amber
+                        Student</a>
                     <br>
                     For University living please refer the link- <a
-                        href="https://www.universityliving.com/australia/city/melbourne"
-                        target="_blank">University Living</a>
+                        href="https://www.universityliving.com/australia/city/melbourne" target="_blank">University
+                        Living</a>
                     <br>
                     For Student Housing please refer the link- <a href="https://sha.com.au/" target="_blank">Student
                         Housing</a>
