+++ model/mdl.js
@@ -1,0 +1,11 @@
+let client = require('../dbConnection');
+
+let feedbackcollection;
+
+feedbackcollection = client.db().collection('Feedback');
+
+function postFeedback(feedback, callback) {
+    feedbackcollection.insertOne(feedback, callback);
+}
+
+module.exports = postFeedback;
